[FIX] Migration: Fix chart migration after 18.5.1
The current introduction of the carousels came required a change of
data structure regarding the charts, namely, split their id from the one
of their container.

Some migration tools (repairInitialMessages) were adapted accordingly
but there were some mmistakes in the adaptation:
- the condition set on the versions was inverted
- the test was adapted but only acounted for the versions that came
  after the refactoring, hiding the previous issue
- the handler  `fixChartDefinitions` did not account for chart
  definitions existing in the carousels

The handler `fixChartDefinitions` was added to level partial definitions
in old chart related commands that were issued around version 16.0
This means that if a user were to create a model with data that come from
version 18.5.1, the data necessarily have been upgraded (snapshotted)
and it is therefore guaranteed that the changes of those commands have
been incorporated in the loaded data, which means the
`fixChartDefinition` no longer makes sense.

This revision fixes the aforementioned issues by skipping `fixChartDefinitions`
for data above 18.5.1. The tests have also been adapted to ensure
previous versions are still properly adapted.

Task: 5365954
X-original-commit: 45725c97603618299f62dd28ba2a697747021cb8
